Add Shawarma recipe and plan To-Do in README.md
Added the Shawarma recipe with clear descriptions, source and steps as well descriptions in code. The code should be organized, so overall it should be finished. But because of the late time, there might be some grammar or spelling mistakes - README.md got updated with a To Do list, to keep plans in track.