Fix list
Fixxed the lists as some of the contents in the lists weren't in the <li></li> tag, even though some browers fix this, some may not. Also fixxed the home anchors as it enclosed a h2 tag but it should be the other way around, HTML5 fixxes this but it isn't cleaner like that.